api rename change: Template -> Render
When trying to name the objects in play with this lib, I was running
into mental roadblocks with the current setup.  You need two objects
in play here: one to store the configuration from the user and another
to be the scope that the template source is evaluated against.

I had originally thought I would use `Template` for the config and
_something_ else for the scope.  But as I tried to name the something
I realized the naming was off.  I think `Template` should be the
scope your source file evals against.  It's called a "template file"
after all.

So here I'm thinking the config object should be called a `Render`.
A render builds and renders a template in the correct format.  The
naming seems to flow better with the action and other names involved.

This does the trivial rename of `Template` to `Render`.