remove "magic underscore" behavior in the `partial` method
This switches partial rendering to not magically include a leading
underscore in the file name.  It is simpler and clearer to just make
the user specify the exact file name and if they choose to use the
convention of leading underscores then whatever.

The result is the `partial` method is now just an alias to calling
`render`.