default template locals
This updates the source to optionally take default template locals.
Locals are applied to an anonymous subclass of `Template` and this
subclass is used to do all source renderings.

The goal here is to apply a local that is used across all renders -
one you don't have to remember to pass in and applies to both top
level renders and partial renders.  Specifically, this will be used
by both the deas and sanford nm adapters to make their `logger`
available in the templates.