template cache
This adds logic to the source to enable template caching.  This
shows a modest improvement in the benchmarks relative to not caching.

To accomplish this, I had to update the source api to take options
rather than default locals directly.  This also updates the the
README to document the cache option as well as the default locals
option.