(hotfix) rename `file_name` usage to `template_name`
This is prep for globbing for source files and removing the requirement
that source files end in `.nm`.  Plus this info might include path
info relative to the source root so 'template_name' is more appropriate
anyway.

I chose to do this as hotfix as I don't want this "noise" in coming
PRs.