don't require that templates end in a `.nm` extension
At this low-level, nm shouldn't be in the business of demanding
that template files end in any one extension.  If Nm is asked to
render a named template, it should just render it.

Note: the motivation for this change came from updating Deas to
support multi-engine extensions on template files.  Via deas-nm
(and also sanford-nm too), developers can configure nm to render
templates for any extension they choose - Nm should honor that.