source: allow optionally specifying an extension for looking up source files
This helps when you have two files names the same thing but with
different extensions.  Previously, only the first file globbed would
be rendered.  This allows you to prefer rendering one over the other.

This scenario should have been covered in PR 16 but was missed.

Note: this also changes to have Nm complain if asked to render a named
template that doesn't exist (whether specifying an extension or not).
This makes for more friendly errors in this scenario.