(hotfix) fix Source file globbing to work in newer ruby versions
When globbing for template source files, Nm checks if the given
template name ends in the Source's configured extension (which
can be `nil` if no extension has been configured).  If the name
ends with the extension, Source globs the given name exactly; if
not, it fuzzy-globs the name.

So in 1.8.7, you can call `String#end_with?` with a `nil` value
and it returns `false` as expected.  In newer versions, calling
that method with a `nil` value raises a `TypeError: no implicit
conversion of nil into String` exception (ugh).

So, this switches to now babysit modern Ruby, hold its hand and
manually handle globbing where there is no configured extension.
This gets things passing in modern Ruby.