rework Source to use keyword arguments; tweak argument handling
This is just some minor updates to modernize how arguments are
passed to the Source object. The only significant change is I've
renamed the `.ext` reader to `.extension` to formally spell out the
property name. I did this to be less cryptic in the API which is
our modern convention when possible/practical.