add `render_content` method to Nm::Context
This method is used to render raw template content strings and is
composed by the `render` method that takes a template path. This
keeps things DRY while adding a more flexible API.

Specifically, this is needed to implement render Nm templates in
Rails' template engine.

I also added some equality methods to ease testing in the rails-nm
gem that I'm working on.